Add Proportions

Add proportions of expstudy metric variables. Other variables can be provided via the <dynamic-dots> argument. Proportions will sum to 100% per group if .base_grp_nms are provided. If .base_grp_nms are omitted, proportions will sum to 100% in total.

Usage

add_proportions(expstudy, ..., .base_grp_nms = character(0), .min_ungrpd = 0L)

Arguments

expstudy

an expstudy()

...

other columns other than the expstudy metric variables (actuals, expecteds, and exposures) to generate proportions of

.base_grp_nms

character vector of column names to use as the base of added proportions

.min_ungrpd

minimum number of non-grouping columns required before using .base_grp_nms as proportion base

Value

An expstudy with added proportions.

Examples

  es <- expstudy(
    data = mortexp,
    actuals = ACTUAL_DEATHS,
    expecteds = EXPECTED_DEATHS,
    exposures =  EXPOSURE,
    variances = VARIANCE_DEATHS
  )

  # If no arguments are provided, proportions will be generated for expstudy
  # metric variables only. This mostly makes sense for already aggregated
  # expstudy objects, but can be used with unaggregated objects as well.
  es %>%
    aggregate(
      ATTAINED_AGE
    ) %>%
    add_proportions
#> Source: local data table [83 x 9]
#> Call:   `_DT3`[, .(ACTUAL_DEATHS = sum(ACTUAL_DEATHS), EXPECTED_DEATHS = sum(EXPECTED_DEATHS), 
#>     EXPOSURE = sum(EXPOSURE), VARIANCE_DEATHS = sum(VARIANCE_DEATHS)), 
#>     keyby = .(ATTAINED_AGE)][, `:=`(PROP_ACTUAL_DEATHS = ACTUAL_DEATHS/sum(ACTUAL_DEATHS), 
#>     PROP_EXPECTED_DEATHS = EXPECTED_DEATHS/sum(EXPECTED_DEATHS), 
#>     PROP_EXPOSURE = EXPOSURE/sum(EXPOSURE), PROP_VARIANCE_DEATHS = VARIANCE_DEATHS/sum(VARIANCE_DEATHS))]
#> 
#>   ATTAINED_AGE ACTUAL_…¹ EXPEC…² EXPOS…³ VARIA…⁴ PROP_…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸
#>          <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1           19         0   0.214    21.7   0.214       0 8.51e-4 0.00152 8.51e-4
#> 2           20         0   0.565    56.5   0.565       0 2.24e-3 0.00396 2.24e-3
#> 3           21         0   0.662    65.5   0.661       0 2.63e-3 0.00459 2.63e-3
#> 4           22         0   0.713    69.9   0.712       0 2.83e-3 0.00490 2.83e-3
#> 5           23         0   0.828    80.3   0.827       0 3.28e-3 0.00563 3.29e-3
#> 6           24         0   0.884    84.9   0.884       0 3.51e-3 0.00595 3.51e-3
#> # … with 77 more rows, and abbreviated variable names ¹​ACTUAL_DEATHS,
#> #   ²​EXPECTED_DEATHS, ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​PROP_ACTUAL_DEATHS,
#> #   ⁶​PROP_EXPECTED_DEATHS, ⁷​PROP_EXPOSURE, ⁸​PROP_VARIANCE_DEATHS
#> 
#> # Use as.data.table()/as.data.frame()/as_tibble() to access results

  # For grouped proportions, use `.base_grp_nms` to identify which variable
  # to use as the base for proportions.
  es %>%
    aggregate(
      GENDER,
      SMOKING_STATUS
    ) %>%
    add_proportions(
      .base_grp_nms = 'GENDER'
    )
#> Source: local data table [4 x 10]
#> Groups: GENDER
#> Call:   `_DT3`[, .(ACTUAL_DEATHS = sum(ACTUAL_DEATHS), EXPECTED_DEATHS = sum(EXPECTED_DEATHS), 
#>     EXPOSURE = sum(EXPOSURE), VARIANCE_DEATHS = sum(VARIANCE_DEATHS)), 
#>     keyby = .(GENDER, SMOKING_STATUS)][, `:=`(PROP_ACTUAL_DEATHS = ACTUAL_DEATHS/sum(ACTUAL_DEATHS), 
#>     PROP_EXPECTED_DEATHS = EXPECTED_DEATHS/sum(EXPECTED_DEATHS), 
#>     PROP_EXPOSURE = EXPOSURE/sum(EXPOSURE), PROP_VARIANCE_DEATHS = VARIANCE_DEATHS/sum(VARIANCE_DEATHS)), 
#>     by = .(GENDER)]
#> 
#>   GENDER SMOKI…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#>   <fct>  <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE NON-SM…      95    76.3   4294.    76.2   0.709   0.712   0.704   0.712
#> 2 FEMALE SMOKER       39    30.9   1804.    30.8   0.291   0.288   0.296   0.288
#> 3 MALE   NON-SM…     146   102.    5662.   102.    0.679   0.707   0.693   0.707
#> 4 MALE   SMOKER       69    42.5   2504.    42.5   0.321   0.293   0.307   0.293
#> # … with abbreviated variable names ¹​SMOKING_STATUS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> # Use as.data.table()/as.data.frame()/as_tibble() to access results

  # `.min_ungrpd` is useful only when generating multiple combinations of
  # results with compile_results(); this prevents proportions all equaling
  # 100% when all grouping columns are used.