Compile resulting analyses

For a streamlined approach, this function combines multiple other expstudy-package functions to generate multiple analyses simultaneously. All grouping combinations will be generated from variables passed as <dynamic-dots>, then analytics will be created for each.

Resulting output can have the typical metrics via add_metrics(), or proportions via add_proportions(). Default parameters are used for each function, however, you can pass additional parameters in a list to the corresponding function element via the output_args argument.

Furthermore, you can choose to have the results return unformatted (for further calculations or analysis) or formatted (for presentation purposes).

This function was meant to provide quick results for routine analysis. Any additional in-depth analyses should make use of individual expstudy functions instead.

Usage

compile_results(
  expstudy,
  ...,
  output = c("metrics", "proportions"),
  output_args = list(metrics = NULL, proportions = NULL),
  output_format = c("unformatted", "formatted")
)

Arguments

expstudy

an expstudy()

...

variables to generate grouped analyses. All combinations of variables provided will be generated and used as groups in results.

output

type of result to output (i.e., with metrics or proportions). If omitted, both will be generated and returned in a list.

output_args

additional arguments to pass to add_metrics() and add_proportions(). If omitted, default parameters will be used for both functions.

output_format

resulting output format. If omitted, both will be generated and returned as a list.

Value

a (potentially nested) list with resulting analysis according to arguments passed to the function.

See also

Examples

  es <- expstudy(
    data = mortexp,
    actuals = ACTUAL_DEATHS,
    expecteds = EXPECTED_DEATHS,
    exposures =  EXPOSURE,
    variances = VARIANCE_DEATHS
  )

  # Quickly generate multiple analyses by simply declaring variables of
  # interest within an experience study.
  es %>%
    compile_results(
      GENDER,
      SMOKING_STATUS,
      UNDERWRITING_CLASS
    )
#> $UNFORMATTED
#> $UNFORMATTED$METRICS
#> $UNFORMATTED$METRICS$AGGREGATE
#> # A tibble: 1 × 7
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS ACTUA…¹ ACTUA…² EXPEC…³
#> *         <dbl>           <dbl>    <dbl>           <dbl>   <dbl>   <dbl>   <dbl>
#> 1           349            252.   14264.            252.    1.38  0.0245  0.0177
#> # … with abbreviated variable names ¹​ACTUAL_TO_EXPECTED, ²​ACTUAL_TO_EXPOSED,
#> #   ³​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY GENDER`
#> # A tibble: 2 × 8
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIAN…¹ ACTUA…² ACTUA…³ EXPEC…⁴
#> * <fct>          <dbl>           <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE           134            107.    6098.     107.    1.25  0.0220  0.0176
#> 2 MALE             215            145.    8166.     145.    1.48  0.0263  0.0177
#> # … with abbreviated variable names ¹​VARIANCE_DEATHS, ²​ACTUAL_TO_EXPECTED,
#> #   ³​ACTUAL_TO_EXPOSED, ⁴​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 8
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTE…¹ EXPOS…² VARIA…³ ACTUA…⁴ ACTUA…⁵ EXPEC…⁶
#> * <fct>                  <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 NON-SMOKER               241     179.    9956.   178.     1.35  0.0242  0.0180
#> 2 SMOKER                   108      73.4   4309.    73.3    1.47  0.0251  0.0170
#> # … with abbreviated variable names ¹​EXPECTED_DEATHS, ²​EXPOSURE,
#> #   ³​VARIANCE_DEATHS, ⁴​ACTUAL_TO_EXPECTED, ⁵​ACTUAL_TO_EXPOSED,
#> #   ⁶​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 8
#>   UNDERWRITING_CLASS ACTUAL_DE…¹ EXPEC…² EXPOS…³ VARIA…⁴ ACTUA…⁵ ACTUA…⁶ EXPEC…⁷
#> * <fct>                    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 PREFERRED                   36    25.7   1579.    25.7    1.40  0.0228  0.0163
#> 2 SELECT                      80    63.0   3534.    62.8    1.27  0.0226  0.0178
#> 3 STANDARD                   233   163.    9151.   163.     1.43  0.0255  0.0179
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​ACTUAL_TO_EXPECTED, ⁶​ACTUAL_TO_EXPOSED,
#> #   ⁷​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 9
#>   GENDER SMOKING_STATUS ACTUAL…¹ EXPEC…² EXPOS…³ VARIA…⁴ ACTUA…⁵ ACTUA…⁶ EXPEC…⁷
#> * <fct>  <fct>             <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE NON-SMOKER           95    76.3   4294.    76.2    1.24  0.0221  0.0178
#> 2 FEMALE SMOKER               39    30.9   1804.    30.8    1.26  0.0216  0.0171
#> 3 MALE   NON-SMOKER          146   102.    5662.   102.     1.43  0.0258  0.0181
#> 4 MALE   SMOKER               69    42.5   2504.    42.5    1.62  0.0276  0.0170
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​ACTUAL_TO_EXPECTED, ⁶​ACTUAL_TO_EXPOSED,
#> #   ⁷​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   GENDER UNDERWRITING_…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷ EXPEC…⁸
#> * <fct>  <fct>             <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE PREFERRED            12    10.6    626.    10.6    1.13  0.0192  0.0169
#> 2 FEMALE SELECT               30    28.4   1602.    28.3    1.06  0.0187  0.0177
#> 3 FEMALE STANDARD             92    68.2   3870.    68.1    1.35  0.0238  0.0176
#> 4 MALE   PREFERRED            24    15.1    953.    15.1    1.59  0.0252  0.0159
#> 5 MALE   SELECT               50    34.6   1932.    34.5    1.45  0.0259  0.0179
#> 6 MALE   STANDARD            141    95.2   5281.    95.0    1.48  0.0267  0.0180
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED,
#> #   ⁷​ACTUAL_TO_EXPOSED, ⁸​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷ EXPEC…⁸
#> * <fct>          <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 NON-SMOKER     PREFER…      26   18.4    1134.   18.4     1.41  0.0229  0.0162
#> 2 NON-SMOKER     SELECT       54   44.9    2453.   44.8     1.20  0.0220  0.0183
#> 3 NON-SMOKER     STANDA…     161  115.     6368.  115.      1.39  0.0253  0.0181
#> 4 SMOKER         PREFER…      10    7.34    445.    7.33    1.36  0.0225  0.0165
#> 5 SMOKER         SELECT       26   18.1    1081.   18.1     1.44  0.0240  0.0167
#> 6 SMOKER         STANDA…      72   48.0    2782.   47.9     1.50  0.0259  0.0172
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED,
#> #   ⁷​ACTUAL_TO_EXPOSED, ⁸​EXPECTED_TO_EXPOSED
#> 
#> $UNFORMATTED$METRICS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 10
#>    GENDER SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷
#>  * <fct>  <fct>          <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 FEMALE NON-SMOKER     PREFER…       9    7.74    455.    7.73    1.16  0.0198
#>  2 FEMALE NON-SMOKER     SELECT       22   20.7    1149.   20.7     1.06  0.0191
#>  3 FEMALE NON-SMOKER     STANDA…      64   47.9    2689.   47.8     1.34  0.0238
#>  4 FEMALE SMOKER         PREFER…       3    2.86    171.    2.86    1.05  0.0175
#>  5 FEMALE SMOKER         SELECT        8    7.68    453.    7.67    1.04  0.0177
#>  6 FEMALE SMOKER         STANDA…      28   20.3    1181.   20.3     1.38  0.0237
#>  7 MALE   NON-SMOKER     PREFER…      17   10.7     679.   10.6     1.60  0.0250
#>  8 MALE   NON-SMOKER     SELECT       32   24.2    1304.   24.1     1.32  0.0245
#>  9 MALE   NON-SMOKER     STANDA…      97   67.6    3679.   67.4     1.44  0.0264
#> 10 MALE   SMOKER         PREFER…       7    4.48    274.    4.47    1.56  0.0256
#> 11 MALE   SMOKER         SELECT       18   10.4     629.   10.4     1.73  0.0286
#> 12 MALE   SMOKER         STANDA…      44   27.6    1602.   27.6     1.59  0.0275
#> # … with 1 more variable: EXPECTED_TO_EXPOSED <dbl>, and abbreviated variable
#> #   names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS, ³​EXPECTED_DEATHS, ⁴​EXPOSURE,
#> #   ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED, ⁷​ACTUAL_TO_EXPOSED
#> 
#> 
#> $UNFORMATTED$PROPORTIONS
#> $UNFORMATTED$PROPORTIONS$AGGREGATE
#> # A tibble: 1 × 8
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIA…¹ PROP_…² PROP_…³ PROP_…⁴ PROP_…⁵
#> *         <dbl>           <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1           349            252.   14264.    252.       1       1       1       1
#> # … with abbreviated variable names ¹​VARIANCE_DEATHS, ²​PROP_ACTUAL_DEATHS,
#> #   ³​PROP_EXPECTED_DEATHS, ⁴​PROP_EXPOSURE, ⁵​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER`
#> # A tibble: 2 × 9
#>   GENDER ACTUAL_DEATHS EXPECTE…¹ EXPOS…² VARIA…³ PROP_…⁴ PROP_…⁵ PROP_…⁶ PROP_…⁷
#> * <fct>          <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE           134      107.   6098.    107.   0.384   0.425   0.428   0.425
#> 2 MALE             215      145.   8166.    145.   0.616   0.575   0.572   0.575
#> # … with abbreviated variable names ¹​EXPECTED_DEATHS, ²​EXPOSURE,
#> #   ³​VARIANCE_DEATHS, ⁴​PROP_ACTUAL_DEATHS, ⁵​PROP_EXPECTED_DEATHS,
#> #   ⁶​PROP_EXPOSURE, ⁷​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 9
#>   SMOKING_STATUS ACTUA…¹ EXPEC…² EXPOS…³ VARIA…⁴ PROP_…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸
#> * <fct>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 NON-SMOKER         241   179.    9956.   178.    0.691   0.709   0.698   0.709
#> 2 SMOKER             108    73.4   4309.    73.3   0.309   0.291   0.302   0.291
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​PROP_ACTUAL_DEATHS, ⁶​PROP_EXPECTED_DEATHS,
#> #   ⁷​PROP_EXPOSURE, ⁸​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 9
#>   UNDERWRITING…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 PREFERRED           36    25.7   1579.    25.7   0.103   0.102   0.111   0.102
#> 2 SELECT              80    63.0   3534.    62.8   0.229   0.250   0.248   0.250
#> 3 STANDARD           233   163.    9151.   163.    0.668   0.648   0.642   0.648
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 10
#>   GENDER SMOKI…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>  <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE NON-SM…      95    76.3   4294.    76.2   0.272   0.303   0.301   0.303
#> 2 FEMALE SMOKER       39    30.9   1804.    30.8   0.112   0.122   0.127   0.122
#> 3 MALE   NON-SM…     146   102.    5662.   102.    0.418   0.406   0.397   0.406
#> 4 MALE   SMOKER       69    42.5   2504.    42.5   0.198   0.169   0.176   0.169
#> # … with abbreviated variable names ¹​SMOKING_STATUS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   GENDER UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>  <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 FEMALE PREFER…      12    10.6    626.    10.6  0.0344  0.0421  0.0439  0.0421
#> 2 FEMALE SELECT       30    28.4   1602.    28.3  0.0860  0.113   0.112   0.113 
#> 3 FEMALE STANDA…      92    68.2   3870.    68.1  0.264   0.271   0.271   0.271 
#> 4 MALE   PREFER…      24    15.1    953.    15.1  0.0688  0.0600  0.0668  0.0600
#> 5 MALE   SELECT       50    34.6   1932.    34.5  0.143   0.137   0.135   0.137 
#> 6 MALE   STANDA…     141    95.2   5281.    95.0  0.404   0.378   0.370   0.378 
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $UNFORMATTED$PROPORTIONS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸
#> * <fct>          <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 NON-SMOKER     PREFER…      26   18.4    1134.   18.4   0.0745  0.0730  0.0795
#> 2 NON-SMOKER     SELECT       54   44.9    2453.   44.8   0.155   0.178   0.172 
#> 3 NON-SMOKER     STANDA…     161  115.     6368.  115.    0.461   0.458   0.446 
#> 4 SMOKER         PREFER…      10    7.34    445.    7.33  0.0287  0.0291  0.0312
#> 5 SMOKER         SELECT       26   18.1    1081.   18.1   0.0745  0.0717  0.0758
#> 6 SMOKER         STANDA…      72   48.0    2782.   47.9   0.206   0.190   0.195 
#> # … with 1 more variable: PROP_VARIANCE_DEATHS <dbl>, and abbreviated variable
#> #   names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS, ³​EXPECTED_DEATHS, ⁴​EXPOSURE,
#> #   ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS, ⁷​PROP_EXPECTED_DEATHS,
#> #   ⁸​PROP_EXPOSURE
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 11
#>    GENDER SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷
#>  * <fct>  <fct>          <fct>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 FEMALE NON-SMOKER     PREFER…       9    7.74    455.    7.73 0.0258   0.0307
#>  2 FEMALE NON-SMOKER     SELECT       22   20.7    1149.   20.7  0.0630   0.0821
#>  3 FEMALE NON-SMOKER     STANDA…      64   47.9    2689.   47.8  0.183    0.190 
#>  4 FEMALE SMOKER         PREFER…       3    2.86    171.    2.86 0.00860  0.0114
#>  5 FEMALE SMOKER         SELECT        8    7.68    453.    7.67 0.0229   0.0305
#>  6 FEMALE SMOKER         STANDA…      28   20.3    1181.   20.3  0.0802   0.0806
#>  7 MALE   NON-SMOKER     PREFER…      17   10.7     679.   10.6  0.0487   0.0423
#>  8 MALE   NON-SMOKER     SELECT       32   24.2    1304.   24.1  0.0917   0.0959
#>  9 MALE   NON-SMOKER     STANDA…      97   67.6    3679.   67.4  0.278    0.268 
#> 10 MALE   SMOKER         PREFER…       7    4.48    274.    4.47 0.0201   0.0178
#> 11 MALE   SMOKER         SELECT       18   10.4     629.   10.4  0.0516   0.0413
#> 12 MALE   SMOKER         STANDA…      44   27.6    1602.   27.6  0.126    0.110 
#> # … with 2 more variables: PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>, and
#> #   abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS
#> 
#> 
#> 
#> $FORMATTED
#> $FORMATTED$METRICS
#> $FORMATTED$METRICS$AGGREGATE
#> # A tibble: 1 × 7
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE  VARIANCE_DEA…¹ ACTUA…² ACTUA…³ EXPEC…⁴
#> * <chr>         <chr>           <chr>     <chr>          <chr>   <chr>   <chr>  
#> 1 349.00        252.12          14,264.29 251.66         138.43% 2.45%   1.77%  
#> # … with abbreviated variable names ¹​VARIANCE_DEATHS, ²​ACTUAL_TO_EXPECTED,
#> #   ³​ACTUAL_TO_EXPOSED, ⁴​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY GENDER`
#> # A tibble: 2 × 8
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIAN…¹ ACTUA…² ACTUA…³ EXPEC…⁴
#> * <fct>  <chr>         <chr>           <chr>    <chr>    <chr>   <chr>   <chr>  
#> 1 FEMALE 134.00        107.18          6,098.01 106.99   125.02% 2.2%    1.76%  
#> 2 MALE   215.00        144.93          8,166.29 144.67   148.35% 2.63%   1.77%  
#> # … with abbreviated variable names ¹​VARIANCE_DEATHS, ²​ACTUAL_TO_EXPECTED,
#> #   ³​ACTUAL_TO_EXPOSED, ⁴​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 8
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTE…¹ EXPOS…² VARIA…³ ACTUA…⁴ ACTUA…⁵ EXPEC…⁶
#> * <fct>          <chr>         <chr>     <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 NON-SMOKER     241.00        178.72    9,955.… 178.39  134.85% 2.42%   1.8%   
#> 2 SMOKER         108.00        73.40     4,308.… 73.27   147.15% 2.51%   1.7%   
#> # … with abbreviated variable names ¹​EXPECTED_DEATHS, ²​EXPOSURE,
#> #   ³​VARIANCE_DEATHS, ⁴​ACTUAL_TO_EXPECTED, ⁵​ACTUAL_TO_EXPOSED,
#> #   ⁶​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 8
#>   UNDERWRITING_CLASS ACTUAL_DE…¹ EXPEC…² EXPOS…³ VARIA…⁴ ACTUA…⁵ ACTUA…⁶ EXPEC…⁷
#> * <fct>              <chr>       <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 PREFERRED          36.00       25.74   1,579.… 25.70   139.87% 2.28%   1.63%  
#> 2 SELECT             80.00       62.95   3,534.… 62.84   127.08% 2.26%   1.78%  
#> 3 STANDARD           233.00      163.43  9,150.… 163.12  142.57% 2.55%   1.79%  
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​ACTUAL_TO_EXPECTED, ⁶​ACTUAL_TO_EXPOSED,
#> #   ⁷​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 9
#>   GENDER SMOKING_STATUS ACTUAL…¹ EXPEC…² EXPOS…³ VARIA…⁴ ACTUA…⁵ ACTUA…⁶ EXPEC…⁷
#> * <fct>  <fct>          <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FEMALE NON-SMOKER     95.00    76.32   4,293.… 76.18   124.48% 2.21%   1.78%  
#> 2 FEMALE SMOKER         39.00    30.87   1,804.… 30.82   126.34% 2.16%   1.71%  
#> 3 MALE   NON-SMOKER     146.00   102.40  5,662.… 102.22  142.57% 2.58%   1.81%  
#> 4 MALE   SMOKER         69.00    42.53   2,504.… 42.45   162.25% 2.76%   1.7%   
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​ACTUAL_TO_EXPECTED, ⁶​ACTUAL_TO_EXPOSED,
#> #   ⁷​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   GENDER UNDERWRITING_…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷ EXPEC…⁸
#> * <fct>  <fct>           <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FEMALE PREFERRED       12.00   10.60   626.27  10.59   113.17% 1.92%   1.69%  
#> 2 FEMALE SELECT          30.00   28.38   1,601.… 28.33   105.71% 1.87%   1.77%  
#> 3 FEMALE STANDARD        92.00   68.20   3,869.… 68.08   134.89% 2.38%   1.76%  
#> 4 MALE   PREFERRED       24.00   15.14   952.86  15.11   158.57% 2.52%   1.59%  
#> 5 MALE   SELECT          50.00   34.57   1,932.… 34.51   144.63% 2.59%   1.79%  
#> 6 MALE   STANDARD        141.00  95.23   5,280.… 95.05   148.07% 2.67%   1.8%   
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED,
#> #   ⁷​ACTUAL_TO_EXPOSED, ⁸​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷ EXPEC…⁸
#> * <fct>          <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 NON-SMOKER     PREFER… 26.00   18.40   1,134.… 18.37   141.32% 2.29%   1.62%  
#> 2 NON-SMOKER     SELECT  54.00   44.87   2,452.… 44.79   120.35% 2.2%    1.83%  
#> 3 NON-SMOKER     STANDA… 161.00  115.45  6,368.… 115.24  139.45% 2.53%   1.81%  
#> 4 SMOKER         PREFER… 10.00   7.34    444.95  7.33    136.24% 2.25%   1.65%  
#> 5 SMOKER         SELECT  26.00   18.08   1,081.… 18.05   143.78% 2.4%    1.67%  
#> 6 SMOKER         STANDA… 72.00   47.97   2,782.… 47.89   150.08% 2.59%   1.72%  
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED,
#> #   ⁷​ACTUAL_TO_EXPOSED, ⁸​EXPECTED_TO_EXPOSED
#> 
#> $FORMATTED$METRICS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 10
#>    GENDER SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ ACTUA…⁶ ACTUA…⁷
#>  * <fct>  <fct>          <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 FEMALE NON-SMOKER     PREFER… 9.00    7.74    454.93  7.73    116.27% 1.98%  
#>  2 FEMALE NON-SMOKER     SELECT  22.00   20.70   1,149.… 20.66   106.29% 1.91%  
#>  3 FEMALE NON-SMOKER     STANDA… 64.00   47.88   2,689.… 47.79   133.68% 2.38%  
#>  4 FEMALE SMOKER         PREFER… 3.00    2.86    171.33  2.86    104.82% 1.75%  
#>  5 FEMALE SMOKER         SELECT  8.00    7.68    452.50  7.67    104.16% 1.77%  
#>  6 FEMALE SMOKER         STANDA… 28.00   20.33   1,180.… 20.29   137.76% 2.37%  
#>  7 MALE   NON-SMOKER     PREFER… 17.00   10.66   679.24  10.64   159.52% 2.5%   
#>  8 MALE   NON-SMOKER     SELECT  32.00   24.17   1,303.… 24.13   132.4%  2.45%  
#>  9 MALE   NON-SMOKER     STANDA… 97.00   67.58   3,679.… 67.45   143.54% 2.64%  
#> 10 MALE   SMOKER         PREFER… 7.00    4.48    273.62  4.47    156.32% 2.56%  
#> 11 MALE   SMOKER         SELECT  18.00   10.40   628.93  10.39   173.04% 2.86%  
#> 12 MALE   SMOKER         STANDA… 44.00   27.65   1,601.… 27.60   159.15% 2.75%  
#> # … with 1 more variable: EXPECTED_TO_EXPOSED <chr>, and abbreviated variable
#> #   names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS, ³​EXPECTED_DEATHS, ⁴​EXPOSURE,
#> #   ⁵​VARIANCE_DEATHS, ⁶​ACTUAL_TO_EXPECTED, ⁷​ACTUAL_TO_EXPOSED
#> 
#> 
#> $FORMATTED$PROPORTIONS
#> $FORMATTED$PROPORTIONS$AGGREGATE
#> # A tibble: 1 × 8
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIA…¹ PROP_…² PROP_…³ PROP_…⁴ PROP_…⁵
#> * <chr>         <chr>           <chr>    <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 349.00        252.12          14,264.… 251.66  100%    100%    100%    100%   
#> # … with abbreviated variable names ¹​VARIANCE_DEATHS, ²​PROP_ACTUAL_DEATHS,
#> #   ³​PROP_EXPECTED_DEATHS, ⁴​PROP_EXPOSURE, ⁵​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER`
#> # A tibble: 2 × 9
#>   GENDER ACTUAL_DEATHS EXPECTE…¹ EXPOS…² VARIA…³ PROP_…⁴ PROP_…⁵ PROP_…⁶ PROP_…⁷
#> * <fct>  <chr>         <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FEMALE 134.00        107.18    6,098.… 106.99  38.4%   42.51%  42.75%  42.52% 
#> 2 MALE   215.00        144.93    8,166.… 144.67  61.6%   57.49%  57.25%  57.48% 
#> # … with abbreviated variable names ¹​EXPECTED_DEATHS, ²​EXPOSURE,
#> #   ³​VARIANCE_DEATHS, ⁴​PROP_ACTUAL_DEATHS, ⁵​PROP_EXPECTED_DEATHS,
#> #   ⁶​PROP_EXPOSURE, ⁷​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 9
#>   SMOKING_STATUS ACTUA…¹ EXPEC…² EXPOS…³ VARIA…⁴ PROP_…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸
#> * <fct>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 NON-SMOKER     241.00  178.72  9,955.… 178.39  69.05%  70.89%  69.79%  70.89% 
#> 2 SMOKER         108.00  73.40   4,308.… 73.27   30.95%  29.11%  30.21%  29.11% 
#> # … with abbreviated variable names ¹​ACTUAL_DEATHS, ²​EXPECTED_DEATHS,
#> #   ³​EXPOSURE, ⁴​VARIANCE_DEATHS, ⁵​PROP_ACTUAL_DEATHS, ⁶​PROP_EXPECTED_DEATHS,
#> #   ⁷​PROP_EXPOSURE, ⁸​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 9
#>   UNDERWRITING…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 PREFERRED      36.00   25.74   1,579.… 25.70   10.32%  10.21%  11.07%  10.21% 
#> 2 SELECT         80.00   62.95   3,534.… 62.84   22.92%  24.97%  24.78%  24.97% 
#> 3 STANDARD       233.00  163.43  9,150.… 163.12  66.76%  64.82%  64.15%  64.82% 
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 10
#>   GENDER SMOKI…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>  <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FEMALE NON-SM… 95.00   76.32   4,293.… 76.18   27.22%  30.27%  30.1%   30.27% 
#> 2 FEMALE SMOKER  39.00   30.87   1,804.… 30.82   11.17%  12.24%  12.65%  12.25% 
#> 3 MALE   NON-SM… 146.00  102.40  5,662.… 102.22  41.83%  40.62%  39.69%  40.62% 
#> 4 MALE   SMOKER  69.00   42.53   2,504.… 42.45   19.77%  16.87%  17.56%  16.87% 
#> # … with abbreviated variable names ¹​SMOKING_STATUS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   GENDER UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸ PROP_…⁹
#> * <fct>  <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FEMALE PREFER… 12.00   10.60   626.27  10.59   3.44%   4.21%   4.39%   4.21%  
#> 2 FEMALE SELECT  30.00   28.38   1,601.… 28.33   8.6%    11.26%  11.23%  11.26% 
#> 3 FEMALE STANDA… 92.00   68.20   3,869.… 68.08   26.36%  27.05%  27.13%  27.05% 
#> 4 MALE   PREFER… 24.00   15.14   952.86  15.11   6.88%   6%      6.68%   6%     
#> 5 MALE   SELECT  50.00   34.57   1,932.… 34.51   14.33%  13.71%  13.55%  13.71% 
#> 6 MALE   STANDA… 141.00  95.23   5,280.… 95.05   40.4%   37.77%  37.02%  37.77% 
#> # … with abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS, ⁸​PROP_EXPOSURE, ⁹​PROP_VARIANCE_DEATHS
#> 
#> $FORMATTED$PROPORTIONS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷ PROP_…⁸
#> * <fct>          <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 NON-SMOKER     PREFER… 26.00   18.40   1,134.… 18.37   7.45%   7.3%    7.95%  
#> 2 NON-SMOKER     SELECT  54.00   44.87   2,452.… 44.79   15.47%  17.8%   17.2%  
#> 3 NON-SMOKER     STANDA… 161.00  115.45  6,368.… 115.24  46.13%  45.79%  44.65% 
#> 4 SMOKER         PREFER… 10.00   7.34    444.95  7.33    2.87%   2.91%   3.12%  
#> 5 SMOKER         SELECT  26.00   18.08   1,081.… 18.05   7.45%   7.17%   7.58%  
#> 6 SMOKER         STANDA… 72.00   47.97   2,782.… 47.89   20.63%  19.03%  19.51% 
#> # … with 1 more variable: PROP_VARIANCE_DEATHS <chr>, and abbreviated variable
#> #   names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS, ³​EXPECTED_DEATHS, ⁴​EXPOSURE,
#> #   ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS, ⁷​PROP_EXPECTED_DEATHS,
#> #   ⁸​PROP_EXPOSURE
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 11
#>    GENDER SMOKING_STATUS UNDER…¹ ACTUA…² EXPEC…³ EXPOS…⁴ VARIA…⁵ PROP_…⁶ PROP_…⁷
#>  * <fct>  <fct>          <fct>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 FEMALE NON-SMOKER     PREFER… 9.00    7.74    454.93  7.73    2.58%   3.07%  
#>  2 FEMALE NON-SMOKER     SELECT  22.00   20.70   1,149.… 20.66   6.3%    8.21%  
#>  3 FEMALE NON-SMOKER     STANDA… 64.00   47.88   2,689.… 47.79   18.34%  18.99% 
#>  4 FEMALE SMOKER         PREFER… 3.00    2.86    171.33  2.86    0.86%   1.14%  
#>  5 FEMALE SMOKER         SELECT  8.00    7.68    452.50  7.67    2.29%   3.05%  
#>  6 FEMALE SMOKER         STANDA… 28.00   20.33   1,180.… 20.29   8.02%   8.06%  
#>  7 MALE   NON-SMOKER     PREFER… 17.00   10.66   679.24  10.64   4.87%   4.23%  
#>  8 MALE   NON-SMOKER     SELECT  32.00   24.17   1,303.… 24.13   9.17%   9.59%  
#>  9 MALE   NON-SMOKER     STANDA… 97.00   67.58   3,679.… 67.45   27.79%  26.8%  
#> 10 MALE   SMOKER         PREFER… 7.00    4.48    273.62  4.47    2.01%   1.78%  
#> 11 MALE   SMOKER         SELECT  18.00   10.40   628.93  10.39   5.16%   4.13%  
#> 12 MALE   SMOKER         STANDA… 44.00   27.65   1,601.… 27.60   12.61%  10.97% 
#> # … with 2 more variables: PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>, and
#> #   abbreviated variable names ¹​UNDERWRITING_CLASS, ²​ACTUAL_DEATHS,
#> #   ³​EXPECTED_DEATHS, ⁴​EXPOSURE, ⁵​VARIANCE_DEATHS, ⁶​PROP_ACTUAL_DEATHS,
#> #   ⁷​PROP_EXPECTED_DEATHS
#> 
#> 
#>