Compile resulting analyses

For a streamlined approach, this function combines multiple other expstudy-package functions to generate multiple analyses simultaneously. All grouping combinations will be generated from variables passed as <dynamic-dots>, then analytics will be created for each.

Resulting output can have the typical metrics via add_metrics(), or proportions via add_proportions(). Default parameters are used for each function, however, you can pass additional parameters in a list to the corresponding function element via the output_args argument.

Furthermore, you can choose to have the results return unformatted (for further calculations or analysis) or formatted (for presentation purposes).

This function was meant to provide quick results for routine analysis. Any additional in-depth analyses should make use of individual expstudy functions instead.

Usage

compile_results(
  expstudy,
  ...,
  output = c("metrics", "proportions"),
  output_args = list(metrics = NULL, proportions = NULL),
  output_format = c("unformatted", "formatted")
)

Arguments

expstudy

an expstudy()

...

variables to generate grouped analyses. All combinations of variables provided will be generated and used as groups in results.

output

type of result to output (i.e., with metrics or proportions). If omitted, both will be generated and returned in a list.

output_args

additional arguments to pass to add_metrics() and add_proportions(). If omitted, default parameters will be used for both functions.

output_format

resulting output format. If omitted, both will be generated and returned as a list.

Value

a (potentially nested) list with resulting analysis according to arguments passed to the function.

See also

Examples

  es <- expstudy(
    data = mortexp,
    actuals = ACTUAL_DEATHS,
    expecteds = EXPECTED_DEATHS,
    exposures =  EXPOSURE,
    variances = VARIANCE_DEATHS
  )

  # Quickly generate multiple analyses by simply declaring variables of
  # interest within an experience study.
  es %>%
    compile_results(
      GENDER,
      SMOKING_STATUS,
      UNDERWRITING_CLASS
    )
#> $UNFORMATTED
#> $UNFORMATTED$METRICS
#> $UNFORMATTED$METRICS$AGGREGATE
#> # A tibble: 1 × 7
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS ACTUAL_TO_EXPECTED
#> *         <dbl>           <dbl>    <dbl>           <dbl>              <dbl>
#> 1           349            252.   14264.            252.               1.38
#> # ℹ 2 more variables: ACTUAL_TO_EXPOSED <dbl>, EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY GENDER`
#> # A tibble: 2 × 8
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>          <dbl>           <dbl>    <dbl>           <dbl>
#> 1 FEMALE           134            107.    6098.            107.
#> 2 MALE             215            145.    8166.            145.
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <dbl>, ACTUAL_TO_EXPOSED <dbl>,
#> #   EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 8
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>                  <dbl>           <dbl>    <dbl>           <dbl>
#> 1 NON-SMOKER               241           179.     9956.           178. 
#> 2 SMOKER                   108            73.4    4309.            73.3
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <dbl>, ACTUAL_TO_EXPOSED <dbl>,
#> #   EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 8
#>   UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>                      <dbl>           <dbl>    <dbl>           <dbl>
#> 1 PREFERRED                     36            25.7    1579.            25.7
#> 2 SELECT                        80            63.0    3534.            62.8
#> 3 STANDARD                     233           163.     9151.           163. 
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <dbl>, ACTUAL_TO_EXPOSED <dbl>,
#> #   EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 9
#>   GENDER SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <fct>                  <dbl>           <dbl>    <dbl>           <dbl>
#> 1 FEMALE NON-SMOKER                95            76.3    4294.            76.2
#> 2 FEMALE SMOKER                    39            30.9    1804.            30.8
#> 3 MALE   NON-SMOKER               146           102.     5662.           102. 
#> 4 MALE   SMOKER                    69            42.5    2504.            42.5
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <dbl>, ACTUAL_TO_EXPOSED <dbl>,
#> #   EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   GENDER UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>  <fct>                      <dbl>           <dbl>    <dbl>
#> 1 FEMALE PREFERRED                     12            10.6     626.
#> 2 FEMALE SELECT                        30            28.4    1602.
#> 3 FEMALE STANDARD                      92            68.2    3870.
#> 4 MALE   PREFERRED                     24            15.1     953.
#> 5 MALE   SELECT                        50            34.6    1932.
#> 6 MALE   STANDARD                     141            95.2    5281.
#> # ℹ 4 more variables: VARIANCE_DEATHS <dbl>, ACTUAL_TO_EXPECTED <dbl>,
#> #   ACTUAL_TO_EXPOSED <dbl>, EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>          <fct>                      <dbl>           <dbl>    <dbl>
#> 1 NON-SMOKER     PREFERRED                     26           18.4     1134.
#> 2 NON-SMOKER     SELECT                        54           44.9     2453.
#> 3 NON-SMOKER     STANDARD                     161          115.      6368.
#> 4 SMOKER         PREFERRED                     10            7.34     445.
#> 5 SMOKER         SELECT                        26           18.1     1081.
#> 6 SMOKER         STANDARD                      72           48.0     2782.
#> # ℹ 4 more variables: VARIANCE_DEATHS <dbl>, ACTUAL_TO_EXPECTED <dbl>,
#> #   ACTUAL_TO_EXPOSED <dbl>, EXPECTED_TO_EXPOSED <dbl>
#> 
#> $UNFORMATTED$METRICS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 10
#>    GENDER SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS
#>  * <fct>  <fct>          <fct>                      <dbl>           <dbl>
#>  1 FEMALE NON-SMOKER     PREFERRED                      9            7.74
#>  2 FEMALE NON-SMOKER     SELECT                        22           20.7 
#>  3 FEMALE NON-SMOKER     STANDARD                      64           47.9 
#>  4 FEMALE SMOKER         PREFERRED                      3            2.86
#>  5 FEMALE SMOKER         SELECT                         8            7.68
#>  6 FEMALE SMOKER         STANDARD                      28           20.3 
#>  7 MALE   NON-SMOKER     PREFERRED                     17           10.7 
#>  8 MALE   NON-SMOKER     SELECT                        32           24.2 
#>  9 MALE   NON-SMOKER     STANDARD                      97           67.6 
#> 10 MALE   SMOKER         PREFERRED                      7            4.48
#> 11 MALE   SMOKER         SELECT                        18           10.4 
#> 12 MALE   SMOKER         STANDARD                      44           27.6 
#> # ℹ 5 more variables: EXPOSURE <dbl>, VARIANCE_DEATHS <dbl>,
#> #   ACTUAL_TO_EXPECTED <dbl>, ACTUAL_TO_EXPOSED <dbl>,
#> #   EXPECTED_TO_EXPOSED <dbl>
#> 
#> 
#> $UNFORMATTED$PROPORTIONS
#> $UNFORMATTED$PROPORTIONS$AGGREGATE
#> # A tibble: 1 × 8
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS PROP_ACTUAL_DEATHS
#> *         <dbl>           <dbl>    <dbl>           <dbl>              <dbl>
#> 1           349            252.   14264.            252.                  1
#> # ℹ 3 more variables: PROP_EXPECTED_DEATHS <dbl>, PROP_EXPOSURE <dbl>,
#> #   PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER`
#> # A tibble: 2 × 9
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>          <dbl>           <dbl>    <dbl>           <dbl>
#> 1 FEMALE           134            107.    6098.            107.
#> 2 MALE             215            145.    8166.            145.
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <dbl>, PROP_EXPECTED_DEATHS <dbl>,
#> #   PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 9
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>                  <dbl>           <dbl>    <dbl>           <dbl>
#> 1 NON-SMOKER               241           179.     9956.           178. 
#> 2 SMOKER                   108            73.4    4309.            73.3
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <dbl>, PROP_EXPECTED_DEATHS <dbl>,
#> #   PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 9
#>   UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>                      <dbl>           <dbl>    <dbl>           <dbl>
#> 1 PREFERRED                     36            25.7    1579.            25.7
#> 2 SELECT                        80            63.0    3534.            62.8
#> 3 STANDARD                     233           163.     9151.           163. 
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <dbl>, PROP_EXPECTED_DEATHS <dbl>,
#> #   PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 10
#>   GENDER SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <fct>                  <dbl>           <dbl>    <dbl>           <dbl>
#> 1 FEMALE NON-SMOKER                95            76.3    4294.            76.2
#> 2 FEMALE SMOKER                    39            30.9    1804.            30.8
#> 3 MALE   NON-SMOKER               146           102.     5662.           102. 
#> 4 MALE   SMOKER                    69            42.5    2504.            42.5
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <dbl>, PROP_EXPECTED_DEATHS <dbl>,
#> #   PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   GENDER UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>  <fct>                      <dbl>           <dbl>    <dbl>
#> 1 FEMALE PREFERRED                     12            10.6     626.
#> 2 FEMALE SELECT                        30            28.4    1602.
#> 3 FEMALE STANDARD                      92            68.2    3870.
#> 4 MALE   PREFERRED                     24            15.1     953.
#> 5 MALE   SELECT                        50            34.6    1932.
#> 6 MALE   STANDARD                     141            95.2    5281.
#> # ℹ 5 more variables: VARIANCE_DEATHS <dbl>, PROP_ACTUAL_DEATHS <dbl>,
#> #   PROP_EXPECTED_DEATHS <dbl>, PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>          <fct>                      <dbl>           <dbl>    <dbl>
#> 1 NON-SMOKER     PREFERRED                     26           18.4     1134.
#> 2 NON-SMOKER     SELECT                        54           44.9     2453.
#> 3 NON-SMOKER     STANDARD                     161          115.      6368.
#> 4 SMOKER         PREFERRED                     10            7.34     445.
#> 5 SMOKER         SELECT                        26           18.1     1081.
#> 6 SMOKER         STANDARD                      72           48.0     2782.
#> # ℹ 5 more variables: VARIANCE_DEATHS <dbl>, PROP_ACTUAL_DEATHS <dbl>,
#> #   PROP_EXPECTED_DEATHS <dbl>, PROP_EXPOSURE <dbl>, PROP_VARIANCE_DEATHS <dbl>
#> 
#> $UNFORMATTED$PROPORTIONS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 11
#>    GENDER SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS
#>  * <fct>  <fct>          <fct>                      <dbl>           <dbl>
#>  1 FEMALE NON-SMOKER     PREFERRED                      9            7.74
#>  2 FEMALE NON-SMOKER     SELECT                        22           20.7 
#>  3 FEMALE NON-SMOKER     STANDARD                      64           47.9 
#>  4 FEMALE SMOKER         PREFERRED                      3            2.86
#>  5 FEMALE SMOKER         SELECT                         8            7.68
#>  6 FEMALE SMOKER         STANDARD                      28           20.3 
#>  7 MALE   NON-SMOKER     PREFERRED                     17           10.7 
#>  8 MALE   NON-SMOKER     SELECT                        32           24.2 
#>  9 MALE   NON-SMOKER     STANDARD                      97           67.6 
#> 10 MALE   SMOKER         PREFERRED                      7            4.48
#> 11 MALE   SMOKER         SELECT                        18           10.4 
#> 12 MALE   SMOKER         STANDARD                      44           27.6 
#> # ℹ 6 more variables: EXPOSURE <dbl>, VARIANCE_DEATHS <dbl>,
#> #   PROP_ACTUAL_DEATHS <dbl>, PROP_EXPECTED_DEATHS <dbl>, PROP_EXPOSURE <dbl>,
#> #   PROP_VARIANCE_DEATHS <dbl>
#> 
#> 
#> 
#> $FORMATTED
#> $FORMATTED$METRICS
#> $FORMATTED$METRICS$AGGREGATE
#> # A tibble: 1 × 7
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE  VARIANCE_DEATHS ACTUAL_TO_EXPECTED
#> * <chr>         <chr>           <chr>     <chr>           <chr>             
#> 1 349.00        252.12          14,264.29 251.66          138.43%           
#> # ℹ 2 more variables: ACTUAL_TO_EXPOSED <chr>, EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY GENDER`
#> # A tibble: 2 × 8
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <chr>         <chr>           <chr>    <chr>          
#> 1 FEMALE 134.00        107.18          6,098.01 106.99         
#> 2 MALE   215.00        144.93          8,166.29 144.67         
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <chr>, ACTUAL_TO_EXPOSED <chr>,
#> #   EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 8
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>          <chr>         <chr>           <chr>    <chr>          
#> 1 NON-SMOKER     241.00        178.72          9,955.55 178.39         
#> 2 SMOKER         108.00        73.40           4,308.74 73.27          
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <chr>, ACTUAL_TO_EXPOSED <chr>,
#> #   EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 8
#>   UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>              <chr>         <chr>           <chr>    <chr>          
#> 1 PREFERRED          36.00         25.74           1,579.12 25.70          
#> 2 SELECT             80.00         62.95           3,534.39 62.84          
#> 3 STANDARD           233.00        163.43          9,150.78 163.12         
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <chr>, ACTUAL_TO_EXPOSED <chr>,
#> #   EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 9
#>   GENDER SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <fct>          <chr>         <chr>           <chr>    <chr>          
#> 1 FEMALE NON-SMOKER     95.00         76.32           4,293.55 76.18          
#> 2 FEMALE SMOKER         39.00         30.87           1,804.46 30.82          
#> 3 MALE   NON-SMOKER     146.00        102.40          5,662.01 102.22         
#> 4 MALE   SMOKER         69.00         42.53           2,504.28 42.45          
#> # ℹ 3 more variables: ACTUAL_TO_EXPECTED <chr>, ACTUAL_TO_EXPOSED <chr>,
#> #   EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   GENDER UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>  <fct>              <chr>         <chr>           <chr>   
#> 1 FEMALE PREFERRED          12.00         10.60           626.27  
#> 2 FEMALE SELECT             30.00         28.38           1,601.94
#> 3 FEMALE STANDARD           92.00         68.20           3,869.80
#> 4 MALE   PREFERRED          24.00         15.14           952.86  
#> 5 MALE   SELECT             50.00         34.57           1,932.45
#> 6 MALE   STANDARD           141.00        95.23           5,280.98
#> # ℹ 4 more variables: VARIANCE_DEATHS <chr>, ACTUAL_TO_EXPECTED <chr>,
#> #   ACTUAL_TO_EXPOSED <chr>, EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 9
#>   SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>          <fct>              <chr>         <chr>           <chr>   
#> 1 NON-SMOKER     PREFERRED          26.00         18.40           1,134.17
#> 2 NON-SMOKER     SELECT             54.00         44.87           2,452.96
#> 3 NON-SMOKER     STANDARD           161.00        115.45          6,368.42
#> 4 SMOKER         PREFERRED          10.00         7.34            444.95  
#> 5 SMOKER         SELECT             26.00         18.08           1,081.44
#> 6 SMOKER         STANDARD           72.00         47.97           2,782.35
#> # ℹ 4 more variables: VARIANCE_DEATHS <chr>, ACTUAL_TO_EXPECTED <chr>,
#> #   ACTUAL_TO_EXPOSED <chr>, EXPECTED_TO_EXPOSED <chr>
#> 
#> $FORMATTED$METRICS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 10
#>    GENDER SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS
#>  * <fct>  <fct>          <fct>              <chr>         <chr>          
#>  1 FEMALE NON-SMOKER     PREFERRED          9.00          7.74           
#>  2 FEMALE NON-SMOKER     SELECT             22.00         20.70          
#>  3 FEMALE NON-SMOKER     STANDARD           64.00         47.88          
#>  4 FEMALE SMOKER         PREFERRED          3.00          2.86           
#>  5 FEMALE SMOKER         SELECT             8.00          7.68           
#>  6 FEMALE SMOKER         STANDARD           28.00         20.33          
#>  7 MALE   NON-SMOKER     PREFERRED          17.00         10.66          
#>  8 MALE   NON-SMOKER     SELECT             32.00         24.17          
#>  9 MALE   NON-SMOKER     STANDARD           97.00         67.58          
#> 10 MALE   SMOKER         PREFERRED          7.00          4.48           
#> 11 MALE   SMOKER         SELECT             18.00         10.40          
#> 12 MALE   SMOKER         STANDARD           44.00         27.65          
#> # ℹ 5 more variables: EXPOSURE <chr>, VARIANCE_DEATHS <chr>,
#> #   ACTUAL_TO_EXPECTED <chr>, ACTUAL_TO_EXPOSED <chr>,
#> #   EXPECTED_TO_EXPOSED <chr>
#> 
#> 
#> $FORMATTED$PROPORTIONS
#> $FORMATTED$PROPORTIONS$AGGREGATE
#> # A tibble: 1 × 8
#>   ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE  VARIANCE_DEATHS PROP_ACTUAL_DEATHS
#> * <chr>         <chr>           <chr>     <chr>           <chr>             
#> 1 349.00        252.12          14,264.29 251.66          100%              
#> # ℹ 3 more variables: PROP_EXPECTED_DEATHS <chr>, PROP_EXPOSURE <chr>,
#> #   PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER`
#> # A tibble: 2 × 9
#>   GENDER ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <chr>         <chr>           <chr>    <chr>          
#> 1 FEMALE 134.00        107.18          6,098.01 106.99         
#> 2 MALE   215.00        144.93          8,166.29 144.67         
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <chr>, PROP_EXPECTED_DEATHS <chr>,
#> #   PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY SMOKING_STATUS`
#> # A tibble: 2 × 9
#>   SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>          <chr>         <chr>           <chr>    <chr>          
#> 1 NON-SMOKER     241.00        178.72          9,955.55 178.39         
#> 2 SMOKER         108.00        73.40           4,308.74 73.27          
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <chr>, PROP_EXPECTED_DEATHS <chr>,
#> #   PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY UNDERWRITING_CLASS`
#> # A tibble: 3 × 9
#>   UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>              <chr>         <chr>           <chr>    <chr>          
#> 1 PREFERRED          36.00         25.74           1,579.12 25.70          
#> 2 SELECT             80.00         62.95           3,534.39 62.84          
#> 3 STANDARD           233.00        163.43          9,150.78 163.12         
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <chr>, PROP_EXPECTED_DEATHS <chr>,
#> #   PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER AND SMOKING_STATUS`
#> # A tibble: 4 × 10
#>   GENDER SMOKING_STATUS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE VARIANCE_DEATHS
#> * <fct>  <fct>          <chr>         <chr>           <chr>    <chr>          
#> 1 FEMALE NON-SMOKER     95.00         76.32           4,293.55 76.18          
#> 2 FEMALE SMOKER         39.00         30.87           1,804.46 30.82          
#> 3 MALE   NON-SMOKER     146.00        102.40          5,662.01 102.22         
#> 4 MALE   SMOKER         69.00         42.53           2,504.28 42.45          
#> # ℹ 4 more variables: PROP_ACTUAL_DEATHS <chr>, PROP_EXPECTED_DEATHS <chr>,
#> #   PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   GENDER UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>  <fct>              <chr>         <chr>           <chr>   
#> 1 FEMALE PREFERRED          12.00         10.60           626.27  
#> 2 FEMALE SELECT             30.00         28.38           1,601.94
#> 3 FEMALE STANDARD           92.00         68.20           3,869.80
#> 4 MALE   PREFERRED          24.00         15.14           952.86  
#> 5 MALE   SELECT             50.00         34.57           1,932.45
#> 6 MALE   STANDARD           141.00        95.23           5,280.98
#> # ℹ 5 more variables: VARIANCE_DEATHS <chr>, PROP_ACTUAL_DEATHS <chr>,
#> #   PROP_EXPECTED_DEATHS <chr>, PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY SMOKING_STATUS AND UNDERWRITING_CLASS`
#> # A tibble: 6 × 10
#>   SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS EXPOSURE
#> * <fct>          <fct>              <chr>         <chr>           <chr>   
#> 1 NON-SMOKER     PREFERRED          26.00         18.40           1,134.17
#> 2 NON-SMOKER     SELECT             54.00         44.87           2,452.96
#> 3 NON-SMOKER     STANDARD           161.00        115.45          6,368.42
#> 4 SMOKER         PREFERRED          10.00         7.34            444.95  
#> 5 SMOKER         SELECT             26.00         18.08           1,081.44
#> 6 SMOKER         STANDARD           72.00         47.97           2,782.35
#> # ℹ 5 more variables: VARIANCE_DEATHS <chr>, PROP_ACTUAL_DEATHS <chr>,
#> #   PROP_EXPECTED_DEATHS <chr>, PROP_EXPOSURE <chr>, PROP_VARIANCE_DEATHS <chr>
#> 
#> $FORMATTED$PROPORTIONS$`BY GENDER, SMOKING_STATUS, AND UNDERWRITING_CLASS`
#> # A tibble: 12 × 11
#>    GENDER SMOKING_STATUS UNDERWRITING_CLASS ACTUAL_DEATHS EXPECTED_DEATHS
#>  * <fct>  <fct>          <fct>              <chr>         <chr>          
#>  1 FEMALE NON-SMOKER     PREFERRED          9.00          7.74           
#>  2 FEMALE NON-SMOKER     SELECT             22.00         20.70          
#>  3 FEMALE NON-SMOKER     STANDARD           64.00         47.88          
#>  4 FEMALE SMOKER         PREFERRED          3.00          2.86           
#>  5 FEMALE SMOKER         SELECT             8.00          7.68           
#>  6 FEMALE SMOKER         STANDARD           28.00         20.33          
#>  7 MALE   NON-SMOKER     PREFERRED          17.00         10.66          
#>  8 MALE   NON-SMOKER     SELECT             32.00         24.17          
#>  9 MALE   NON-SMOKER     STANDARD           97.00         67.58          
#> 10 MALE   SMOKER         PREFERRED          7.00          4.48           
#> 11 MALE   SMOKER         SELECT             18.00         10.40          
#> 12 MALE   SMOKER         STANDARD           44.00         27.65          
#> # ℹ 6 more variables: EXPOSURE <chr>, VARIANCE_DEATHS <chr>,
#> #   PROP_ACTUAL_DEATHS <chr>, PROP_EXPECTED_DEATHS <chr>, PROP_EXPOSURE <chr>,
#> #   PROP_VARIANCE_DEATHS <chr>
#> 
#> 
#>