Mortality experience study

A dataset containing an example of a mortality experience study for 1000 fictional whole life insurance policyholders.

Usage

mortexp

Format

A data.frame() with over 175,000 rows and 24 columns:

AS_OF_DATE

This indicates which point in time a record encompasses.

POLICY_HOLDER

An index used to distinguish policyholders. In this example the policyholder is also the (only) insured.

GENDER, SMOKING_STATUS, UNDERWRITING_CLASS, INSURED_DOB, ISSUE_DATE, ISSUE_AGE

Various characteristics of an insured at time of issue.

FACE_AMOUNT

Face amount of insurance for a corresponding policy.

TERMINATION_DATE

If terminated, the effective date of termination. An NA value will be listed for policies that are still in-force.

ATTAINED_AGE

The age of the insured at the record's AS_OF_DATE

EXPECTED_MORTALITY_RT

An expected mortality rate for an insured. The rate is calculated according to De Moivre's Law (also known as uniform distribution of deaths, or \(\text{UDD}\)) with \(\omega=120\).

POLICY_DURATION_MNTH, POLICY_DURATION_YR

Temporal indices describing how long a policy has been in-force at the AS_OF_DATE. For example, when a policy is first issued (i.e., \(t=0\)), it is in policy duration year one and policy duration month one.

POLICY_STATUS

The current status of the policy, either in-force, surrendered, or death. The value will be listed for each policy record even though a decrement only occurs at the end of the policy's duration (for policies which are no longer in-force).

MORT_EXPOSURE_CNT,MORT_EXPOSURE_AMT

Measures how many policyholders or how much face amount of insurance is exposed to the risk of decrement for an associated observations.

MORT_ACTUAL_CNT,MORT_ACTUAL_AMT

Measures the decrement occurrence on a policy count or face amount of insurance basis.

MORT_EXPECTED_CNT,MORT_EXPECTED_AMT

Measures the expected decrement value for an associated observation on a policy count or face amount of insurance basis.

MORT_VARIANCE_CNT,MORT_VARIANCE_AMT

Measures the variance of the decrement expectation, also on a policy count or face amount of insurance basis. Used to calculate credibility scores and confidence intervals.

Source

All policy record detail is randomly generated. See the Society of Actuaries' publication on experience study calculations for additional information regarding experience study calculations.